make gathering of metrics reusable
